default camera gst and handle Camera args
